# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Rest in Memory
  description: Rest api for in memory key value storage
  contact:
    email: kaanhakan2005@gmail.com
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local environment
tags:
  - name: go-rest-inmemory
    description: Rest api for in memory key value storage.
paths:
  /datas:
    post:
      summary: Creates key value pair
      operationId: createData
      tags:
        - go-rest-inmemory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value:
                    type: string
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
  /datas/{key}:
    get:
      summary: Gets value for key
      operationId: getData
      tags:
        - go-rest-inmemory
      parameters:
        - name: key
          in: path
          description: key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:
                    type: string
                  value:
                    type: string
        '404':
          description: Not Found
        '500':
          description: Internal Server Error
  /datas/flush:
    delete:
      summary: Deletes all key value pairs
      operationId: deleteAllData
      tags:
        - go-rest-inmemory
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Internal Server Error